Modifies handling of stereo samples
Instead of mixing stereo samples to mono and then panning, this
modification uses the `pan` value as a stereo balance, and no mixdown
is performed.

Classic Dirt did not mix down stereo samples unless the "pan" parameter
was present (but did standard mono mixdown if it was), so this change
makes SuperDirt act more like Dirt when "pan" is unspecified.