Experimental sound combos
Basic functionality is there, but very fragile.

Because the Tidal parser doesn't support the notation yet, you need to
use something like:
`d1 $ s (listToPat ["foo^rave"])`.  The current behavior is the
multiply the two together, but this is easy to modify in the SynthDef.
Fragile means it'll probably crash if you try to use something that
isn't a sample or that is undefined.  If a sample is in stereo, only
one channel is read (it'll complain about this).